refactor: restructure GitHub Action codebase by removing unused files and consolidating environment utilities
