+++ packages/github-action/src/environment.ts
@@ -1,0 +1,27 @@
+/**
+ * Environment variable utilities
+ */
+
+export function getGitHubToken(): string {
+  const githubToken = process.env.GITHUB_TOKEN;
+  if (!githubToken) {
+    throw new Error(
+      "GitHub token not found. Please ensure the `github-token` input is set in your workflow.",
+    );
+  }
+  return githubToken;
+}
+
+export function getPochiConfig() {
+  const pochiToken = process.env.POCHI_SESSION_TOKEN;
+  if (!pochiToken) {
+    throw new Error(
+      "POCHI_SESSION_TOKEN environment variable is required. Please add your pochi session token to GitHub Secrets as POCHI_SESSION_TOKEN.",
+    );
+  }
+
+  return {
+    token: pochiToken,
+    model: process.env.POCHI_MODEL || "",
+  };
+}
