--- packages/github-action/src/github/index.ts
@@ -1,7 +1,0 @@
-/**
- * GitHub module exports
- */
-
-export * from "./api";
-export * from "./parser";
-export { GitHubManager } from "./github-manager";
