--- packages/github-action/src/utils/index.ts
@@ -1,6 +1,0 @@
-/**
- * Utility function exports
- */
-
-export * from "./environment";
-export * from "./url";
